refactor: AWS to break out config from functions
